(BSR)[PRO] refactor: remove spec folder
