+++ pro/src/components/Bookings/BookingsRecapTable/BookingsRecapTable.spec.tsx
@@ -16,11 +16,10 @@ import { bookingRecapFactory } from 'commons/utils/factories/individualApiFactor
 import { renderWithProviders } from 'commons/utils/renderWithProviders'
 import * as filterBookingsRecap from 'components/Bookings/BookingsRecapTable/utils/filterBookingsRecap'
 
-import { BookingsRecapTable } from '../BookingsRecapTable'
-import {
-  EMPTY_FILTER_VALUE,
-  DEFAULT_OMNISEARCH_CRITERIA,
-} from '../Filters/constants'
+import { BookingsRecapTable } from './BookingsRecapTable'
+import { EMPTY_FILTER_VALUE, DEFAULT_OMNISEARCH_CRITERIA } from './Filters/constants'
+
+
 
 vi.mock('commons/utils/windowMatchMedia', () => ({
   doesUserPreferReducedMotion: vi.fn(() => true),
